Align token actions menu with the rest of the project
