+++ src/app/index.tsx
@@ -2,7 +2,6 @@ import { StatusCard } from "@/components/status-card";
 import { ThemedText } from "@/components/themed-text";
 import { ThemedView } from "@/components/themed-view";
 import { TokenListItem } from "@/components/token-list-item";
-import { type TokenAction } from "@/components/token-actions-menu";
 import { Radii, Spacing, StaticColors, Typography } from "@/constants/theme";
 import { useChallengePolling } from "@/hooks/use-challenge-polling";
 import { useDeleteTokenConfirmation } from "@/hooks/use-delete-token-confirmation";
@@ -10,6 +9,7 @@ import { useDevMenu } from "@/hooks/use-dev-menu";
 import { useNotificationStatus } from "@/hooks/use-notifications";
 import { useTheme } from "@/hooks/use-theme";
 import { useToken } from "@/hooks/use-token";
+import type { TokenAction } from "@/types/token-actions";
 import type { PushToken } from "@/types/token";
 import { PushTokenRolloutState } from "@/types/token";
 import AddSymbol from "@expo/material-symbols/add.xml";
